@@ -14,33 +14,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.6",
+    "@angular/animations": "^5.0.0",
     "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "^4.3.6",
-    "@angular/compiler": "^4.3.6",
-    "@angular/core": "^4.3.6",
-    "@angular/forms": "^4.3.6",
-    "@angular/http": "^4.3.6",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
     "@angular/material": "^2.0.0-beta.12",
-    "@angular/platform-browser": "^4.3.6",
-    "@angular/platform-browser-dynamic": "^4.3.6",
-    "@angular/router": "^4.3.6",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "angularfire2": "^4.0.0-rc.2",
     "core-js": "^2.5.1",
     "firebase": "^4.3.0",
     "hammerjs": "^2.0.8",
-    "rxjs": "^5.4.3",
+    "rxjs": "^5.5.2",
     "zone.js": "^0.8.17"
   },
   "devDependencies": {
-    "@angular/cli": "^1.3.2",
-    "@angular/compiler-cli": "^4.3.6",
+    "@angular/cli": "^1.5.0",
+    "@angular/compiler-cli": "^5.0.0",
     "@angular/service-worker": "^1.0.0-beta.16",
     "@types/hammerjs": "^2.0.35",
     "@types/jasmine": "2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.88",
-    "codelyzer": "~3.1.1",
+    "codelyzer": "^3.2.2",
     "firebase-tools": "^3.10.0",
     "jasmine-core": "2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
@@ -53,7 +54,7 @@
     "karma-junit-reporter": "^1.2.0",
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "tslint": "^5.7.0",
+    "typescript": "^2.4.2"
   }
 }
